requirements.txt
django==5.0.6
djangorestframework==3.15.1
django-debug-toolbar==4.4.2
django-split-settings==1.3.1
pydantic-settings==2.3.0
pydantic==2.7.2
psycopg-binary==3.1.19
psycopg[binary]==3.1.19
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
loguru==0.7.2
drf-yasg==1.21.7
instaloader==4.11
gunicorn==22.0.0